Added support for configurable rockbox directory. FS#9567 by Alex Bennee.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@19208 a1c6a512-1295-4272-9138-f99709370657

+++ apps/gui/wps_parser.c
@@ -443,7 +443,7 @@ static char *get_image_filename(const char *start, const char* bmpdir,
 {
     const char *end = strchr(start, '|');
 
-    if ( !end || (end - start) >= (buf_size - ROCKBOX_DIR_LEN - 2) )
+    if ( !end || (end - start) >= (buf_size - (int)ROCKBOX_DIR_LEN - 2) )
     {
         buf = "\0";
         return NULL;
